Changed output so it is the integer value rather than float.

+++ combined/main/ICM42688/gyroscope.h
@@ -9,8 +9,8 @@ float gyroscope_set_fs(ICM42688 *imuhdl, uint8_t value);
 float gyroscope_get_odr(ICM42688 *imuhdl);
 float gyroscope_set_odr(ICM42688 *imuhdl, uint8_t value);
 
-float gyroscope_x(ICM42688 *imuhdl);
-float gyroscope_y(ICM42688 *imuhdl);
-float gyroscope_z(ICM42688 *imuhdl);
+int16_t gyroscope_x(ICM42688 *imuhdl);
+int16_t gyroscope_y(ICM42688 *imuhdl);
+int16_t gyroscope_z(ICM42688 *imuhdl);
 
 #endif
